% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hle}
\alias{hle}
\title{Calculate the non parametric Hodges-Lehmann estimator of location for a
\emph{vector} or a single-column \emph{time series} using \code{RcppArmadillo}
and \code{RcppParallel}.}
\usage{
hle(vec_tor)
}
\arguments{
\item{\code{vec_tor}}{A \emph{vector} or a single-column \emph{time series}.}
}
\value{
A single \emph{double} value representing Hodges-Lehmann estimator of 
  the vector.
}
\description{
Calculate the non parametric Hodges-Lehmann estimator of location for a
\emph{vector} or a single-column \emph{time series} using \code{RcppArmadillo}
and \code{RcppParallel}.
}
\details{
The function \code{hle()} calculates the Hodges-Lehmann estimator of 
  the \emph{vector}, using \code{RcppArmadillo} and \code{RcppParallel}. The 
  function \code{hle()} is very much faster than function \code{wilcox.test()}
  in \code{R}.
}
\examples{
\dontrun{
# Create a vector of random returns
re_turns <- rnorm(1e6)
# Compare hle() with wilcox.test()
all.equal(drop(NPE::hle(re_turns)), 
  wilcox.test(re_turns, conf.int = TRUE))
# Compare the speed of RcppParallel with R code
library(microbenchmark)
summary(microbenchmark(
  rcpp=NPE::hle(re_turns),
  rcode=wilcox.test(re_turns, conf.int = TRUE),
  times=10))[, c(1, 4, 5)]  # end microbenchmark summary
}

}
