% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kruskalWalliceTest}
\alias{kruskalWalliceTest}
\title{Performs a Kruskal-Wallis rank sum test. using \code{Rcpp} and \code{boost}.}
\usage{
kruskalWalliceTest(x)
}
\arguments{
\item{\code{x}}{A \emph{List} of numeric data vectors}
}
\value{
A \emph{double} indicating p-value of the test.
}
\description{
Performs a Kruskal-Wallis rank sum test. using \code{Rcpp} and \code{boost}.
}
\details{
The function \code{kruskalWalliceTest()} performs a Kruskal-Wallis rank 
  sum test of the null hypothesis that the location parameters of the distribution
  of x are the same in each group. The alternative is that they differ in
  at least in one.
}
\examples{
\dontrun{
x <- c(2.9, 3.0, 2.5, 2.6, 3.2) # normal subjects
y <- c(3.8, 2.7, 4.0, 2.4)      # with obstructive airway disease
z <- c(2.8, 3.4, 3.7, 2.2, 2.0) # with asbestosis

# Perform Kruskal-Wallis rank sum test on the elements in two ways
all.equal(kruskal.test(list(x, y, z))$p.value, drop(NPE::kruskalWalliceTest(list(x, y, z))))
# Compare the speed of Rcpp and R code
library(microbenchmark)
summary(microbenchmark(
  Rcpp=NPE::kruskalWalliceTest(list(x, y, z)),
  Rcode=kruskal.test(list(x, y, z))$p.value,
  times=10))[, c(1, 4, 5)]  # end microbenchmark summary
}

}
