% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{med_ian}
\alias{med_ian}
\title{Calculate the median of a  \emph{vector} or a single-column \emph{time
series} using \code{RcppArmadillo}.}
\usage{
med_ian(vec_tor)
}
\arguments{
\item{\code{vec_tor}}{A \emph{vector} or a single-column \emph{time series}.}
}
\value{
A single \emph{double} value representing median of the vector.
}
\description{
Calculate the median of a  \emph{vector} or a single-column \emph{time
series} using \code{RcppArmadillo}.
}
\details{
The function \code{med_ian()} calculates the median of the
  \emph{vector}, using \code{RcppArmadillo}. The function \code{med_ian()}
  is several times faster than \code{median()} in \code{R}.
}
\examples{
\dontrun{
# Create a vector of random returns
re_turns <- rnorm(1e6)
# Compare med_ian() with median()
all.equal(drop(NPE::med_ian(re_turns)), 
  median(re_turns))
# Compare the speed of RcppArmadillo with R code
library(microbenchmark)
summary(microbenchmark(
  rcpp=NPE::med_ian(re_turns),
  rcode=median(re_turns),
  times=10))[, c(1, 4, 5)]  # end microbenchmark summary
}

}
