% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{medianAbsoluteDeviation}
\alias{medianAbsoluteDeviation}
\title{Calculate the Median absolute deviation of a  \emph{vector} or a single-column
 \emph{time series} using \code{RcppArmadillo}.}
\usage{
medianAbsoluteDeviation(vec_tor)
}
\arguments{
\item{\code{vec_tor}}{A \emph{vector} or a single-column \emph{time series}.}
}
\value{
A single \emph{double} value representing median absolue deviation of 
  the vector.
}
\description{
Calculate the Median absolute deviation of a  \emph{vector} or a single-column
 \emph{time series} using \code{RcppArmadillo}.
}
\details{
The function \code{medianAbsoluteDeviation()} calculates the median of 
  the \emph{vector}, using \code{RcppArmadillo}. The function \code{medianAbsoluteDeviation()}
  is several times faster than \code{mad()} in \code{R}.
}
\examples{
\dontrun{
# Create a vector of random returns
re_turns <- rnorm(1e6)
# Compare medianAbsoluteDeviation() with mad()
all.equal(drop(HighFreq::medianAbsoluteDeviation(re_turns)), 
  mad(re_turns))
# Compare the speed of RcppArmadillo with R code
library(microbenchmark)
summary(microbenchmark(
  rcpp=HighFreq::medianAbsoluteDeviation(re_turns),
  rcode=mad(re_turns),
  times=10))[, c(1, 4, 5)]  # end microbenchmark summary
}

}
