\name{NPE-package}
\alias{NPE-package}
\alias{NPE}
\docType{package}
\title{
\packageTitle{NPE}
}
\description{
\packageDescription{NPE}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{NPE}
\packageIndices{NPE}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{NPE}

Maintainer: \packageMaintainer{NPE}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
