% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rolling_mad}
\alias{rolling_mad}
\title{Calculate the rolling median absolute deviation over a \emph{vector} or a
single-column \emph{time series} using \code{RcppArmadillo} and
\code{RcppParallel}.}
\usage{
rolling_mad(vec_tor, look_back)
}
\arguments{
\item{\code{vec_tor}}{A \emph{vector} or a single-column \emph{time series}.}

\item{\code{look_back}}{The length of look back interval, equal to the
number of elements of data used for calculating the median.}
}
\value{
A column \emph{vector} of the same length as the argument
  \code{vect_tor}.
}
\description{
Calculate the rolling median absolute deviation over a \emph{vector} or a
single-column \emph{time series} using \code{RcppArmadillo} and
\code{RcppParallel}.
}
\details{
The function \code{rolling_mad()} calculates a vector of
  rolling medians, over a \emph{vector} of data, using \emph{RcppArmadillo}
  and \emph{RcppParallel}.
}
\examples{
\dontrun{
# Create a vector of random returns
re_turns <- rnorm(1e6)
rolling_mad(re_turns)
}

}
