% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rolling_skew}
\alias{rolling_skew}
\title{Calculate the skewness of the columns of a \emph{time series} or a
\emph{matrix} over a rolling window using \code{RcppArmadillo} and
\code{RcppParallel}.}
\usage{
rolling_skew(t_series, look_back, typ_e = "pearson", al_pha = 0.25)
}
\arguments{
\item{\code{t_series}}{A \emph{time series} or a \emph{matrix} of data.}

\item{\code{look_back}}{The length of look back interval.}

\item{\code{typ_e}}{A \emph{string} specifying the type of skewness (see
Details). (The default is the \code{typ_e = "pearson"}.)}

\item{\code{al_pha}}{The confidence level for calculating the quantiles.
(the default is \code{al_pha = 0.25}).}
}
\value{
A matrix with the skewness of the columns of
  \code{t_series} over rolling window.
}
\description{
Calculate the skewness of the columns of a \emph{time series} or a
\emph{matrix} over a rolling window using \code{RcppArmadillo} and
\code{RcppParallel}.
}
\details{
The function \code{rolling_skew()} calculates the skewness of the
  columns of a \emph{time series} or a \emph{matrix} of data using
  \code{RcppArmadillo} and \code{RcppParallel} \code{C++} code.

  If \code{typ_e = "pearson"} (the default) then \code{calc_skew()}
  calculates the Pearson skewness using the third moment of the data.

  If \code{typ_e = "quantile"} then it calculates the skewness using the
  differences between the quantiles of the data.

  If \code{typ_e = "nonparametric"} then it calculates the skewness as the
  difference between the mean of the data minus its median, divided by the
  standard deviation.
  
  The code examples below compare the function \code{rolling_skew()} with the
  skewness calculated using \code{R} code.
}
