% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{theilSenEstimator}
\alias{theilSenEstimator}
\title{Calculate the nonparametric Theil-Sen estimator of dependency-covariance for two
\emph{vectors}  using \code{RcppArmadillo}}
\usage{
theilSenEstimator(x, y)
}
\arguments{
\item{\code{vector_x}}{A \emph{vector} independent (explanatory) data.}

\item{\code{vector_y}}{A \emph{vector} dependent data.}
}
\value{
A column \emph{vector} containing two values i.e intercept and slope
}
\description{
Calculate the nonparametric Theil-Sen estimator of dependency-covariance for two
\emph{vectors}  using \code{RcppArmadillo}
}
\details{
The function \code{theilSenEstimator()} calculates the Theil-Sen
  estimator of the \emph{vector}, using \code{RcppArmadillo} . The function
  \code{theilSenEstimator()} is significantly faster than function
  \code{WRS::tsreg()} in \code{R}.
}
\examples{
\dontrun{
# Create a vector of random returns
vector_x <- rnorm(10)
vactor_y <- rnorm(10)
# Compare theilSenEstimator() with tsreg()
# Compare the speed of RcppParallel with R code
library(microbenchmark)
summary(microbenchmark(
  rcpp=NPE::theilSenEstimator(vector_x, vector_y),
  rcode=WRS(vector_x, vector_y),
  times=10))[, c(1, 4, 5)]  # end microbenchmark summary
}

}
