% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{WilcoxanSignedRankTest}
\alias{WilcoxanSignedRankTest}
\title{Performs one sample Wilcoxan ranked sum test on \emph{vector} or a single-column
\emph{time series} using \code{RcppArmadillo} and \code{boost}.}
\usage{
WilcoxanSignedRankTest(x, mu = 0, alternative = "two.sided",
  exact = FALSE, correct = TRUE)
}
\arguments{
\item{\code{x}}{A \emph{vector} or a single-column \emph{time series}.}

\item{\code{mu}}{A \emph{double} specifing an optional parameter used 
to form null hypothesis. Default value is \emph{zero}.}

\item{\code{alternative}}{a \emph{character} string specifying the alternative
hypothesis. It must be one of :
\itemize{
  \item "two.sided" two tailed test.
  \item "greater" greater(right) tailed test.
  \item "less" smaller(left) tailed test.
}
(The default is \emph{two.sided} test.)}

\item{\code{exact}}{A {boolean} indicating whether an exact p-value should be computed.}

\item{\code{correct}}{A {boolean} indicating whether to apply continuity correction
in normal approximation for the p-value.}
}
\value{
A \emph{double} indicating p-value of the test.
}
\description{
Performs one sample Wilcoxan ranked sum test on \emph{vector} or a single-column
\emph{time series} using \code{RcppArmadillo} and \code{boost}.
}
\details{
The function \code{WilcoxanSignedRankTest()} carries out the wilcoxan signed 
  rank test on \emph{vec_tor} and returns the \emph{p-value} of the test.
  By default (if \code{exact} is not specified), an exact p-value is computed if sample 
  contains less than 50 finite values and there are no ties. Otherwise, a normal approximation
  is used.
}
\examples{
\dontrun{
# Create a vector of random data
da_ta <- round(runif(7), 2)
# Carry out wilcoxan signed rank test on the elements in two ways
all.equal(wilcox.test(da_ta)$p.value, drop(HighFreq::WilcoxanSignedRankTest(da_ta)))
# Create a time series of random data
da_ta <- xts::xts(runif(7), seq.Date(Sys.Date(), by=1, length.out=7))
# Calculate the ranks of the elements in two ways
all.equal(wilcox.test(coredata(da_ta))$p.value, drop(HighFreq::WilcoxanSignedRankTest(da_ta)))
# Compare the speed of Rcpp and R code
da_ta <- runif(10)
library(microbenchmark)
summary(microbenchmark(
  rcpp=WilcoxanSignedRankTest(da_ta),
  rcode=wilcox.test(da_ta),
  times=10))[, c(1, 4, 5)]  # end microbenchmark summary
}

}
