% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{theilSenEstimator}
\alias{theilSenEstimator}
\title{Calculate the nonparametric Theil-Sen estimator of dependency-covariance for
two \emph{vectors} using \code{RcppArmadillo}}
\usage{
theilSenEstimator(x, y)
}
\arguments{
\item{\code{vector_x}}{A \emph{vector} independent (explanatory) data.}

\item{\code{vector_y}}{A \emph{vector} dependent data.}
}
\value{
A column \emph{vector} containing two values i.e intercept and
  slope.
}
\description{
Calculate the nonparametric Theil-Sen estimator of dependency-covariance for
two \emph{vectors} using \code{RcppArmadillo}
}
\details{
The function \code{theilSenEstimator()} calculates the Theil-Sen
  estimator  using \code{RcppArmadillo}. The function
  \code{theilSenEstimator()} is significantly faster than function
  \code{WRS::tsreg()} in \code{R}.
}
\examples{
\dontrun{
# Create vectors of random returns
vector_x <- rnorm(10)
vector_y <- rnorm(10)
# Install package akima and WRS
install.packages("akima")
install.packages("WRS", repos="http://R-Forge.R-project.org")
# Compare theilSenEstimator() with WRS::tsreg()
all.equal(NPE::theilSenEstimator(vector_x, vector_y), 
  WRS::tsreg(vector_x, vector_y, FALSE)$coef, check.attributes=FALSE)
# Compare the speed of RcppParallel with R code
library(microbenchmark)
summary(microbenchmark(
  Rcpp=NPE::theilSenEstimator(vector_x, vector_y),
  Rcode=WRS::tsreg(vector_x, vector_y, FALSE),
  times=10))[, c(1, 4, 5)]  # end microbenchmark summary
}

}
